backend/requirements.txt
# requirements.txt

Flask==2.3.2
ultralytics==8.3.78

# PyTorch ecosystem:
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

numpy==1.26.0
Pillow==10.0.0
